backend/requirements.txt
python-dotenv
ultralytics
fastapi
uvicorn
SQLAlchemy
psycopg2
